backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
scikit-learn
python-multipart
Pillow
joblib